Versona AI/Voice Gen Project/requirements.txt
# Versona AI - Urdu Voice Generation Platform
# Python dependencies

# Core
python-dotenv==1.0.0
requests==2.31.0

# API Server
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Authentication
pyjwt==2.8.0
passlib==1.7.4

# Validation
pydantic==1.10.13
email-validator==2.1.0
